general changes done
